+++ NEWS.md
@@ -5,3 +5,11 @@
 # caretSDM 1.1.1
 
 -   Inclusion of Non-native Distribution using caretSDM with adaptations on sdm_area function to allow the use of scenarios that do not particularly match the modeling area.
+
+# caretSDM 1.1.2
+
+-   Inclusion of methods to include new algorithms in caretSDM.
+
+# caretSDM 1.1.3
+
+-   Inclusion of Ensemble of Small Models in caretSDM.